Ticket Pocket is user-friendly Web Application where user are able to book their movie tickets.
Ticket Pocket is an Web Application which is implemented using HTML Web Language where users are able to book their movie tickets by login to this website, accordingly the amount to be paid will be calculated and displayed for user convenience, at this level user are able to change their desision and get back without confirming the seat and amount. After they confirms number of tickets ,which show timings and amount, they proceed to payment and enjoys their show.